+++ DTO/NewMonthlyCalendarDTO.cs
@@ -6,7 +6,7 @@ public class NewMonthlyCalendarDTO
 {
     public Guid Id { get; set; } = Guid.NewGuid();
 
-    public int Strike { get; set; }
+    public double Strike { get; set; }
 
     public DateOnly ExecutionDate { get; set; } = new DateOnly();
 
@@ -16,9 +16,9 @@ public class NewMonthlyCalendarDTO
 
     public DateOnly BuyOrderExpiryDate { get; set; }
 
-    public double CallSellLTP { get; set; }
-    public double PutSellLTP { get; set; }
-    public double CallBuyLTP { get; set; }
-    public double PutBuyLTP { get; set; }
+    public decimal CallSellLTP { get; set; }
+    public decimal PutSellLTP { get; set; }
+    public decimal CallBuyLTP { get; set; }
+    public decimal PutBuyLTP { get; set; }
     public ClassicalCalendarStatus ClassicalCalendarStatus { get; set; }
 }
